pkg/auth: move auth code into a new package for reuse in syz-hub
